Reorganize support code into ilc.feature.memoize
Also update MemoizeSpec and memoize.ToScala to load the memoize library.

On the way, drop inclusion of maps.SyntaxSugar.
